pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyrox-insights"
version = "0.1.0"
description = "A short description of your project"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Add your dependencies here, e.g., "requests>=2.25.0"
    "pyrox-client (>=0.2.3,<0.3.0)",
    "pandas",
    "dash",
    "plotly",
    "dash-bootstrap-components"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.setuptools]
packages = ["pyrox"]  # Adjust if your package structure differs

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true